+++ tasks/grunt-release.js
@@ -222,14 +222,19 @@ module.exports = function(grunt){
         return;
       }
 
+      if (!process.env[options.github.usernameVar]) {
+        grunt.log.warn('Error: No username for GitHub release');
+      }
+
       request
         .post('https://api.github.com/repos/' + options.github.repo + '/releases')
         .auth(process.env[options.github.usernameVar], process.env[options.github.passwordVar])
         .set('Accept', 'application/vnd.github.manifold-preview')
         .set('User-Agent', 'grunt-release')
         .send({
-          "tag_name": tagName,
-          "name": tagMessage
+          'tag_name': tagName,
+          name: tagMessage,
+          prerelease: type === 'prerelease'
         })
         .end(function(res){
           if (res.statusCode === 201){